Support for Bitwarden API key authentication
Bitwarden servers now require captcha responses for all non-
API clients. Require the use of the API key to bypass this
restriction.

Remove call to /accounts/prelogin as KDF information is
returned in the API key login flow.

API scoped authentication does not provide a refresh token and so
requires a call to /connect/token/ with the API key id & secret
when the access token expires

It may be possible to retrieve the user email via some
API call but that capability is unclear as of now.