make clipboard manipulation happen from the agent
on x11 systems, you can't just send data to the os to store on the
clipboard, you just register which application currently owns the
clipboard and then other applications can use ipc to request the owning
application to send them the clipboard data. this requires there to be
an application still running in order to respond to those requests.
luckily, we have one of those available in the form of the agent.