more correct websocket notification handling
the servers tend to be fairly chatty with messages, mostly pings and
heartbeats of various sorts, and we don't want to sync on all of those.
also, the message type in the first array element of the messagepack
structure is not the same thing as the UpdateType - that is stored as
an argument to the ReceiveMessage invocation, so we need to parse a bit
further to get the actual UpdateType. this still just does a full sync
on any changes, though.