fix rbw-pinentry-keyring not saving passwords correctly
the rbw-pinentry-keyring setup command did not support passwords that includes a space. The cut command only retrieved the first word of the password and thus skipping the other words. `cut -c3-` cuts from the 3rd char 'D ' to the end of the line. 

So pinentry output like 'D this is a passphrase' will be cut to 'this is a passphrase'.