Swap toml_lite for toml_rs and add Steam TOTP
This was discussed privately, but the author of this project wanted for
me to migrate to a seperate TOTP library in order to implement Steam
TOTP key support (as that was my original wish)

Surprisingly not a lot had to be changed in order to accomodate this
library so it seems to be a good choice.

Another small change I made is that I added .idea (from RustRover) to
gitignore because I feel those "project" files are inappropriate to
include in a project like this and they just end up polluting the
repository.