Avoid loading fields that would force a master password reprompt in search
We introduce a new field, `skip_protected` in the `decrypt` request
to the agent. The agent would simply reply with `CipherIsProtected` when asked
to decrypt a cipher that would otherwise force a master password reprompt.

The client would by default bypass all such fields during `get` and `search`
(note that the reprompt would still happen as part of fetching the fields for _displaying_);
unless prompted with `--include-protected-fields`, in which the old behavior will occur
as before.